Added seal functionality
v1 keyserver functionalitiy pending

+++ src/cmd/seal.sh
@@ -21,18 +21,19 @@ cmd_seal() {
 
     # get recipients
     get_gpg_key
-    local recipients="--recipient $GPG_KEY"
+    local recipients="$GPG_KEY"
     while [[ -n "$1" ]]; do
-        recipients="$recipients --recipient $1"
+        recipients="$recipients $1"
         shift
     done
 
     # sign and encrypt
     gnupghome_setup
-    gpg --no-tty --auto-key-locate=local,cert,keyserver,pka \
-        --keyserver "$KEYSERVER" $recipients \
-        --sign --encrypt --armor \
-        --output "$file.sealed" "$file"
+    call_gpg seal.py $file $recipients
+    
+    local err=$?
+    [[ $err == 0 ]] || fail "Error encrypting $file"
+
     gnupghome_reset
 
     [[ -s "$file.sealed" ]] || rm -f "$file.sealed"
